// cms/libs/system-config/domain/src/lib/enums/applicant-info-lov-types.enum.ts
export enum ApplicantInfoLovType {
  EnglishProficiency = "ENGLISH_PROFICIENCY",
  SpokenWrittenLanguage = "SPOKEN_WRITTEN_LANGUAGE" ,
  Gender = "GENDER",
  EthnicityOtherCategories = "ETHNICITY_OTHER_CATEGORIES",
  Transgender = "TRANSGENDER",
  SexAtBirth = "SEX_AT_BIRTH",
  Pronouns = "PRONOUNS",
  SexualIdentity = "SEXUAL_IDENTITY",
  Ethnicity = "ETHNICITY",
  Race = "RACE",
  MaterialInAlternateFormat = "MATERIAL_IN_ALTERNATE_FORMAT"
}